backend/requirements.txt
-r requirements.prod.txt

pep8==1.7.1
autopep8==2.0.4
pycodestyle==2.11.0
uritemplate==4.1.1
celery==5.5.3
django-celery-beat==2.8.1
django-redis==6.0.0